% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-assignments.R
\name{get_artifacts}
\alias{get_artifacts}
\title{List the artifacts for a repo (should be username/repo format)}
\usage{
get_artifacts(repo)
}
\description{
List the artifacts for a repo (should be username/repo format)
}
