% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-assignments.R
\name{get_assignments}
\alias{get_assignments}
\title{Get assignments for a classroom
GH_TOKEN must be allowed access to content}
\usage{
get_assignments(classroom_id)
}
\description{
Get assignments for a classroom
GH_TOKEN must be allowed access to content
}
